Move all availability types and test data to a new file and import.
